% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_wrappers.R
\name{bbox_sfc}
\alias{bbox_sfc}
\title{Get a polygon for a bounding box around an sf object.}
\usage{
bbox_sfc(sf_obj, epsg = 3310)
}
\arguments{
\item{sf_obj}{object of class sf.}

\item{epsg}{is either an integer for the EPSG id or the WKT for defining the coordinate reference system.}
}
\value{
sf object.
}
\description{
Get a polygon for a bounding box around an sf object.
}
