% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdrs-package.R
\docType{package}
\name{cdrs-package}
\alias{cdrs}
\alias{cdrs-package}
\title{cdrs: Functions for the California Delta Residents Survey Data}
\description{
Facilitates the proper loading of California Delta Residents Survey data (2023) available on Open ICPSR, https://doi.org/10.3886/E195447V1. This package includes tools for reading the data into the R environment correctly.
}
\author{
\strong{Maintainer}: Kenji Tomari \email{ktomari@ucdavis.edu} (\href{https://orcid.org/0000-0001-8039-3999}{ORCID})

}
\keyword{internal}
