% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{cdrs_as_NA}
\alias{cdrs_as_NA}
\title{Convert <Missing Levels> to NA}
\usage{
cdrs_as_NA(data_)
}
\arguments{
\item{data_}{A data.frame or tibble consisting of the data from the DRS.}
}
\value{
A data.frame or tibble.
}
\description{
Convert columns with missing values (eg. <I Don't Know>, ie values denoted by <angle brackets>) to \code{NA}.
}
\examples{
df <- data.frame(Q0_0 = c("1", "<Missing>", "3"))
cdrs_as_NA(df)
}
