% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{cdrs_composite_index}
\alias{cdrs_composite_index}
\title{Create composite indices in the 2023 Summary Report.}
\usage{
cdrs_composite_index(data_)
}
\arguments{
\item{data_}{is the full DRS data set (created by \code{cdrs_read}).}
}
\value{
A tibble. the full DRS data set + composite scores.
}
\description{
This function largely reproduces the composite indices created for the Summary Report (albeit with some corrections).
}
