% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_wrappers.R
\name{cdrs_crosstab}
\alias{cdrs_crosstab}
\title{Run a 2-way weighted cross-tabulation on DRS data.}
\usage{
cdrs_crosstab(data_, cols_)
}
\arguments{
\item{data_}{data.frame or tibble, the DRS data set.}

\item{cols_}{a character vector of column names.}
}
\value{
xtabs object from svytable()
}
\description{
Run a 2-way weighted cross-tabulation on DRS data.
}
\examples{
results <- cdrs::cdrs_crosstab(
     data_ = cdrs::cdrs_read_example(),
     cols_ = c("AGE_P", "Q1_1")
     )
}
