% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_wrappers.R
\name{cdrs_design}
\alias{cdrs_design}
\title{Create Survey Design Object.}
\usage{
cdrs_design(data_, set_fpc = F)
}
\arguments{
\item{data_}{A tibble or data.frame which contains the columns/variables of
interest, as well as the Zone and WTFINAL columns which concern the
stratification and weights used in specifying the survey design,
respectively.}

\item{set_fpc}{logical. Determines if we need to set the finite population
correction.}
}
\value{
A R object of class survey.design (from {survey})
}
\description{
This function wraps around survey::svydesign() to appropriately create a
survey design object.
}
\section{Notice}{


The authors of this package are not subject matter experts in the field of
survey statistics, or statistics more broadly. Caution should be taken when
interpreting our notes, and when using our code. If you encounter any
errors, please contact the package maintainer.
}

\section{Conceptual Background}{


Like most surveys, the DRS has a complex sampling design, which involves
non-response and self-selection error. In order to mitigate bias, we need to
utilize "weights" that increase or decrease the importance of each
respondent's contribution to estimates of the population or any given
statistic. While it is not difficult to incorporate weights in calculating
frequencies, survey::svydesign greatly reduces the burden of calculating
variances. Moreover, we use survey::svydesign to ease documentation and
reproducibility. For details on how weights were constructed please visit
the project documentation.
}

\examples{
df <- data.frame(
  Q0_0 = c("1", NA, "3"),
  Zone = c(1, 1, 2),
  WTFINAL = c(1.32, 0.83, 0.98)
)
cdrs_design(df)
}
