% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_wrappers.R
\name{cdrs_gt_simple}
\alias{cdrs_gt_simple}
\title{Simple Crosstabulation Combined Table}
\usage{
cdrs_gt_simple(prep_)
}
\arguments{
\item{prep_}{list. Output of cdrs_gt_prep}
}
\value{
gt object.
}
\description{
Calculates a contingency table showing weighted counts and row percentages as a \{gt\} table.
}
