% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_wrappers.R
\name{cdrs_hist_table}
\alias{cdrs_hist_table}
\title{Calculates cdrs numeric variables as a partitioned histogram.}
\usage{
cdrs_hist_table(data_, col_, return_tb = TRUE)
}
\arguments{
\item{data_}{tibble. The full DRS data set.}

\item{col_}{character. The column name of the numeric variable, eg. "Q1a".}

\item{return_tb}{logical. Whether to return the object returned by \code{survey::svyhist}, or to return a nicely formatted tibble (ready for plotting with ggplot2).}
}
\value{
either an object of class histogram (aka svyhist object), or a tibble.
}
\description{
This function largely serves as a wraparound for \code{survey::svyhist()}. In addition to running \code{svyhist}, this function does two things: offers an output object of class tibble (or a svyhist "histogram" object), and prevents a plot from automatically being generated by \code{svyhist}.
}
