% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{cdrs_labels_table}
\alias{cdrs_labels_table}
\title{Create custom CDRS labels tibble.}
\usage{
cdrs_labels_table()
}
\value{
tibble.
}
\description{
Creates custom CDRS data labels for each question in the public data set. These labels help in making succinct and clear plots. These labels are available as a CSV file in the {cdrs} package itself, under /extdata. This function makes the table (as a tibble) easily accessible.
}
