% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_map_county}
\alias{cdrs_map_county}
\title{Main function to create ggplot2 map on TIGER county geometry.}
\usage{
cdrs_map_county(prep_)
}
\arguments{
\item{prep_}{is a list derived from \code{cdrs_map_prep}.}
}
\description{
Main function to create ggplot2 map on TIGER county geometry.
}
