% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_map_prep}
\alias{cdrs_map_prep}
\title{Map data preparer}
\usage{
cdrs_map_prep(
  data_,
  col_,
  level_ = NA_character_,
  geo_var = "geoid.county",
  param_file = system.file("extdata", "plot_parameters.xlsx", package = "cdrs")
)
}
\arguments{
\item{data_}{is the DRS data set.}

\item{col_}{is a character vector specifying the column to prepare.}

\item{level_}{is a character (or integer, more on this later) of length 1 that specifies a factor level for which to calculate the mean (eg. we want the mean for just "Urban" or just "Suburban" residents). A \code{level_} \emph{must} be specified for categorical variables. If an integer (eg. as either \code{1L} or \code{1}), the level corresponding that number will be selected. For example, 1 for Q2 would select "Urban".}

\item{geo_var}{is a character vector, length 1.}

\item{param_file}{is a file path to a plot_parameters xlsx file.}
}
\value{
list with type, proportions, column name of interest, level name of interest (if applicable), and the geographic variable name.
}
\description{
Prepares data to create a ggplot2 plot with spatial features.
}
