% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_pie}
\alias{cdrs_pie}
\title{Plot a pie chart.}
\usage{
cdrs_pie(data_, col_, weighted_ = TRUE, remove_angle_brackets = TRUE)
}
\arguments{
\item{data_}{the DRS data containing at least the column specified in the next parameter. If weighted is TRUE, both Zone and WTFINAL should also be available in data_.}

\item{col_}{column name (as a character)}

\item{weighted_}{should we use weighted data?}

\item{remove_angle_brackets}{logical. Should we remove brackets around \if{html}{\out{<missingness>}} variables?}
}
\value{
an object of class ggplot.
}
\description{
Plot a pie chart.
}
\examples{
dat <- cdrs_read_example()
cdrs_pie(dat, "Q2")
}
