% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_bar}
\alias{cdrs_plt_bar}
\title{Dichotomous Question Bar Plots.}
\usage{
cdrs_plt_bar(data_, cols_, level_)
}
\arguments{
\item{data_}{the DRS data containing at least the column specified in the next parameter. If weighted is TRUE, both Zone and WTFINAL should also be available in data_.}

\item{cols_}{column name(s) (as a character vector)}

\item{level_}{character.}
}
\value{
an object of class ggplot.
}
\description{
Plot questions from the DRS data set with only one level of interest (eg. questions where we only want to show "Yes" responses).
}
