% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_labels}
\alias{cdrs_plt_labels}
\title{Create labels for plots.}
\usage{
cdrs_plt_labels(dict_, cols_)
}
\arguments{
\item{dict_}{is the uncompromised data dictionary.}

\item{cols_}{variable/column names of the DRS data.}
}
\value{
tibble. A modified subset of the original dictionary.
}
\description{
Choose which kinds of text decoration your plot will have.
}
