% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_pie}
\alias{cdrs_plt_pie}
\title{Plot a pie chart.}
\usage{
cdrs_plt_pie(prep_)
}
\arguments{
\item{prep_}{the tibble returned from \code{cdrs_plt_prep.}}
}
\value{
an object of class ggplot.
}
\description{
Plot a pie chart.
}
