% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_prep}
\alias{cdrs_plt_prep}
\title{Prepare data for plotting.}
\usage{
cdrs_plt_prep(
  data_,
  cols_,
  dict_ = NULL,
  remove_angle_brackets = TRUE,
  is_weighted = TRUE,
  txt_options = NULL,
  sort_ = TRUE,
  title_size = 14,
  label_threshold = 10,
  add_legend_percent = TRUE,
  param_file = system.file("extdata", "plot_parameters.xlsx", package = "cdrs")
)
}
\arguments{
\item{data_}{tibble. the DRS data.}

\item{cols_}{character. A vector of the columns of interest.}

\item{dict_}{tibble. the data dictionary. If \code{NULL} no plot label decoration performed. In other words, the plot will not display textual descriptions.}

\item{remove_angle_brackets}{logical. Determines whether to erase angle brackets from \if{html}{\out{<missingness>}} values.}

\item{is_weighted}{logical.}

\item{txt_options}{list. Either \code{NULL} or a list providing parameters for \code{cdrs_plt_txt()}.}

\item{sort_}{logical. Sort variables/levels by magnitude of the mean.}

\item{title_size}{numeric. The size of the font for the title. All other fonts scale linearly to title_size, even if a title isn't included.}

\item{label_threshold}{numeric. The percent threshold by which we should stop displaying labels (in \code{geom_label}) for stacked bar plots. If no threshold is desired, enter \code{NULL}.}

\item{add_legend_percent}{logical. When a single variable is being plotted with values the \code{label_threshold}, we can choose to show the percentage of the missing labels in the legend which appears below the plot.}

\item{param_file}{character. Path to custom parameters xlsx document. See inst/extdata/plot_parameters.xlsx for file structure.}
}
\value{
a list object, often named \code{prep_} in the \{cdrs\} package.
}
\description{
Prepare DRS data for plotting.
}
