% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_stacked}
\alias{cdrs_plt_stacked}
\title{Creates stacked bar plot.}
\usage{
cdrs_plt_stacked(prep_)
}
\arguments{
\item{prep_}{the tibble returned from \code{cdrs_plt_prep.}}
}
\value{
an object of class ggplot.
}
\description{
Creates stacked bar plot which shows all levels of a variable. Used for both qualitative and likert scale responses.
}
