% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_txt}
\alias{cdrs_plt_txt}
\title{Create text for plots.}
\usage{
cdrs_plt_txt(
  dict_,
  cols_,
  qid_labels = NULL,
  title_ = NULL,
  subtitle_ = FALSE,
  caption_ = FALSE,
  valid_n = NULL,
  param_file = system.file("extdata", "plot_parameters.xlsx", package = "cdrs")
)
}
\arguments{
\item{dict_}{is the data dictionary. All columns in \code{cols_} must be present in \code{dict_$Variable}.}

\item{cols_}{variable/column names of the DRS data.}

\item{qid_labels}{character. Options include "short", "alphabet", \code{NULL}. These labels correspond to the table created by \code{plt_labels()}. If "short", either the variable \code{short_label} or \code{short_lvl} is drawn from the labels table and table for recoding factors accordingly is returned. If "alphabet", a table of for recoding factors aligned with alphabetical characters is returned. If \code{NULL} no table is returned.}

\item{title_}{character. Options include "short", "long" or \code{NULL}. If "short" the \code{short_title} is retrieved from the labels table (see \code{plt_labels()}), If "long", the Label value is retrieved from \code{dict_}. If \code{NULL}, nothing is returned.}

\item{subtitle_}{logical. If \code{TRUE} it returns the long title (ie. Label) from the \code{dict_}.}

\item{caption_}{logical. If \code{TRUE} a detailed message include valid response counts and missingness variable counts returned.}

\item{valid_n}{numeric or \code{NULL}. Not strictly necessary, but this overrides the value returned from dict_ for Valid Responses. This is useful for old versions of the data dictionary that reported incorrect values.}

\item{param_file}{character, path to plot_parameters file.}
}
\value{
list. The list could be empty, or have one or more of the following: 'labels', 'title', 'subtitle', and/or 'caption'.
}
\description{
Create different kinds of text decoration for your plot, including labels that appear along the graph axis, titles, subtitles & captions.
}
