% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{cdrs_read_dict}
\alias{cdrs_read_dict}
\title{Read CDRS data dictionary.}
\usage{
cdrs_read_dict(path_)
}
\arguments{
\item{path_}{is the path to the dictionary.}
}
\value{
tibble.
}
\description{
Helper function that reads the CDRS data dictionary in only.
}
