% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{cdrs_revise}
\alias{cdrs_revise}
\title{Revises DRS data such that different forms of <Missing Levels> may be converted to NA.}
\usage{
cdrs_revise(
  data_,
  dd,
  preserve_uncertainty = T,
  preserve_refused = F,
  preserve_factor_numeric = F,
  preserve_editorials = F
)
}
\arguments{
\item{data_}{A data.frame or tibble consisting of the data from the DRS.}

\item{dd}{Either a path to the zip/metadata folder, OR a data.frame or tibble consisting of the data dictionary (ie. sheet 'Variables' of the data dictionary xlsx file).}

\item{preserve_uncertainty}{logical. This preserves factor levels that indicate some kind of uncertainty for categorical variables (that are not a numeric entry). By default, this is \code{TRUE}. In effect, this preserves levels like \verb{<I don't know>} as is. However, if this parameter is set to \code{FALSE}, levels that signify respondent uncertainty are converted to \code{NA} (ie. R's "Not Available" category). This latter case could be useful if you're primarily concerned with clear affirmatives and negatives. (Note, this parameter does not impact ordinal-likert variables with a "neutral" value.)}

\item{preserve_refused}{logical. This preserves factor levels that indicate that a user read, then intentionally skipped a question, such as \if{html}{\out{<Decline to answer>}}. By default this parameter is \code{FALSE} with the assumption that statistical analysis on the DRS data is employed with the given survey weights. When this parameter is \code{FALSE}, these levels are converted to \code{NA} values (ie. R's "Not Available" category).}

\item{preserve_factor_numeric}{logical. This preserves the non-numeric responses that survey respondents may have had to questions concerning a numeric entry prompt. By default this parameter is \code{FALSE}, such that non-numeric responses like <I don't know> or \if{html}{\out{<Not applicable>}} are converted to \code{NA} (ie. R's "Not Available" category). This facilitates easier estimation of numeric population parameters like the mean.}

\item{preserve_editorials}{logical. There are number of editorialized variables in the DRS data set. These were either created to align with census/marketing categories, like the gender binary, or had values erased for the sake of respondent non-identifiability. By default this value is \code{FALSE}, which thus converts these values to \code{NA} (ie. R's "Not Available" category).}
}
\value{
A data.frame or tibble.
}
\description{
Convert columns with missing values (eg. <I Don't Know>, ie values denoted by <angle brackets>) to \code{NA}. A dependency of cdrs_read().
}
