% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_wrappers.R
\name{cdrs_svytb}
\alias{cdrs_svytb}
\title{Run a 2-way weighted cross-tabulation on DRS data.}
\usage{
cdrs_svytb(
  data_,
  cols_,
  set_fpc = TRUE,
  is_rounded = FALSE,
  is_props = TRUE,
  is_table = FALSE
)
}
\arguments{
\item{data_}{data.frame or tibble, the DRS data set.}

\item{cols_}{a character vector of column names.}

\item{set_fpc}{logical. \code{NULL} defaults to the default of \code{cdrs_design}. See documentation on \code{cdrs_design}.}

\item{is_rounded}{logical. Should we set \code{survey::svytable} argument \code{round} to true?}

\item{is_props}{logical. Whether to return proportions or frequencies (ie. numerical count).}

\item{is_table}{logical. Whether to return the default object of svytable() which is an object of class "svytable", "xtabs", and "table", or to convert it into a data.frame.}
}
\value{
svyby object
}
\description{
Run a 2-way weighted cross-tabulation on DRS data.
}
\examples{
results <- cdrs::cdrs_svytb(
     data_ = cdrs::cdrs_read_example(return_dict = FALSE),
     cols_ = c("AGE_P", "Q1_1")
     )
}
