% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{cdrs_validate}
\alias{cdrs_validate}
\title{Validate DRS data against a SHA256 hash.}
\usage{
cdrs_validate(path_)
}
\arguments{
\item{path_}{is a file path to a directory containing the documents. This should not be a zip file.}
}
\value{
fl, a tibble, including file name, path, and type of data set.
}
\description{
Checks and validates that the data stored in the given directory is valid according to the design of the DRS team. It should have three components, a csv, an excel file, and a txt file. This function is not usually called by the end-user, it is a dependency of cdrs_read.
}
