% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{enrich_dict}
\alias{enrich_dict}
\title{Creates an enriched data dictionary}
\usage{
enrich_dict(dict)
}
\arguments{
\item{dict}{is the DRS data dictionary.}
}
\value{
a tibble of the DRS data dictionary.
}
\description{
Creates a data dictionary where question labels are split into their "prompt" and "response" components. This function heavily relies on tidyr's nesting functions.
}
