% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{generate_grp}
\alias{generate_grp}
\title{Add a grouping column.}
\usage{
generate_grp(dict_, var_ = "Variable")
}
\arguments{
\item{dict_}{the data dictionary}

\item{var_}{the column for the variable ID's.}
}
\value{
tibble of dict_ with new column 'grp'
}
\description{
Add a new column in the data set based on qid grouping variable. This
character column will be filled with values like 1, 1a, 2, 3, and so on.
}
