% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{get_unwt_props}
\alias{get_unwt_props}
\title{Get factor proportions for plotting.}
\usage{
get_unwt_props(data_, col_)
}
\arguments{
\item{data_}{full data set.}

\item{col_}{column name (as a character)}
}
\value{
tibble of unweighted proportions.
}
\description{
Get factor proportions for plotting.
}
\examples{
dat <- cdrs_read_example()
cdrs:::get_unwt_props(dat, "Q2")
}
