% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_wrappers.R
\name{is_dark}
\alias{is_dark}
\title{Determine whether a hex color is dark or light.}
\usage{
is_dark(colors)
}
\arguments{
\item{colors}{character. A hexadecimal color.}
}
\value{
logical. Answers the question, "Is it dark?"
}
\description{
Using ITU Rec. 709 coefficients, we can determine if a color is light or dark for HDTV and other digital screens.
}
\details{
Caution, this implementation has not been verified to correctly determine luminosity. Moreover, this function has not been determined to be ADA compliant.
}
