% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_wrappers.R
\name{map_dichotomous}
\alias{map_dichotomous}
\title{Plot map of dichotmous variable}
\usage{
map_dichotomous(props_sf, is_choropleth = TRUE, limits_ = c(0, 1), k_ = 7L)
}
\arguments{
\item{props_sf}{a sf data.frame}

\item{is_choropleth}{logical}

\item{limits_}{numeric length 2. The range of possible values.}

\item{k_}{integer. Number of classes.}
}
\description{
Plot map of dichotmous variable
}
