% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{matches_to_NA}
\alias{matches_to_NA}
\title{Helper function to convert bracketed factor values to NA.}
\usage{
matches_to_NA(vec_, match_ = NULL)
}
\arguments{
\item{vec_}{a factor vector.}

\item{match_}{a character vector of matching phrases.}
}
\value{
factor vector.
}
\description{
Helper function to convert bracketed factor values to NA.
}
\examples{
tmp <- factor(letters[1:5])
convert_to_NA(tmp, c("a", "d"))
}
