% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_wrappers.R
\name{plt_filter_caps}
\alias{plt_filter_caps}
\title{Filters NA/NULL values out of the 'captions' list.}
\usage{
plt_filter_caps(captions_)
}
\arguments{
\item{captions_}{list. A list with each named element composed of a character, NA, or NULL.}
}
\value{
list.
}
\description{
Filter out NA/NULL values from the 'captions' list in \code{cdrs_plt_prep}.
}
