% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_wrappers.R
\name{plt_legend_nrow}
\alias{plt_legend_nrow}
\title{Determine number of rows for the legend}
\usage{
plt_legend_nrow(title_size, fcts)
}
\arguments{
\item{title_size}{numeric. The font size in points.}

\item{fcts}{character or factor. Each level in the factor variable.}
}
\value{
integer, number of rows.
}
\description{
This function uses a linear model to estimate the number of legend items per row for plots with legends placed on the bottom.
}
