% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{qual_pal}
\alias{qual_pal}
\title{Creates qualitative palette of hex color values.}
\usage{
qual_pal(n_fct, seed = NA)
}
\arguments{
\item{n_fct}{is the number of factors/colors we need.}

\item{seed}{is a numeric value that gives the seed for "random" color
selection within the appropriate palette.}
}
\value{
an palette (character vector).
}
\description{
These palettes were selected based on the greatest minimum distance between all sets of colors in the palette. This distance was calculated using \code{colorblindcheck::palette_check}, which creates a set of scores for each palette for each of the three major types of color vision deficiencies: deuteranopia, protanopia, and tritanopia.
}
\examples{
# Return first 5 colors of Okabe-Ito
cdrs:::qual_pal(5)
}
