% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_wrappers.R
\name{ranked_to_num}
\alias{ranked_to_num}
\title{Convert ordinal scale columns to a numeric score.}
\usage{
ranked_to_num(
  data_,
  col_,
  reverse_ = FALSE,
  zero_start = FALSE,
  convert_missing = TRUE
)
}
\arguments{
\item{data_}{DRS data.frame/tibble.}

\item{col_}{string.}

\item{reverse_}{logical. If we should reverse the levels. By default, the order of factor levels is not reversed because the input factors do not align with the output factors, which \emph{may} cause confusion.}

\item{zero_start}{logical. If we should have the first level be worth 0 or 1. In most cases of ordinal scale factors where the lowest level is something akin to "none", 0 is preferrable.}

\item{convert_missing}{logical. Convert missingness variables to \code{NA}.}
}
\value{
a tibble (the data).
}
\description{
Convert ordinal scale columns to a numeric score.
}
