% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{revise_dict}
\alias{revise_dict}
\title{Revise data dictionary based on changed values in data.}
\usage{
revise_dict(data_, dict_)
}
\arguments{
\item{data_}{revised DRS data.}

\item{dict_}{original data dictionary}
}
\value{
tibble (dict)
}
\description{
Following systematic changes in the data (such as those performed in \code{cdrs_revise}), revise the data dictionary to reflect actual data values.
}
