% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_wrappers.R
\name{cdrs_gt_prep}
\alias{cdrs_gt_prep}
\title{Prepare data for \{gt\} table.}
\usage{
cdrs_gt_prep(
  data_,
  col1,
  col2,
  dict_ = NULL,
  add_labs = TRUE,
  add_title = FALSE,
  label_threshold = 20,
  param_file = system.file("extdata", "plot_parameters.xlsx", package = "cdrs")
)
}
\arguments{
\item{data_}{data.frame. The full CDRS data set.}

\item{col1}{character. The first QID of interest, eg "Q1_1" or "Zone".}

\item{col2}{character. The second QID of interest.}

\item{dict_}{data.frame. The CDRS data dictionary.}

\item{add_labs}{logical. Whether to add labels (ie. for the stubhead and spanner).}

\item{add_title}{logical. Whether to add the gt title and subtitle.}

\item{label_threshold}{numeric. The character at which to wrap labels using \code{stringr::str_wrap()}.}

\item{param_file}{character. Path to plot_parameters.xlsx.}
}
\value{
list (named \code{prep_} in other \{cdrs\} functions).
}
\description{
Obtain metadata needed to create a table with \code{cdrs_gt_simple}.
}
