% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_bar}
\alias{cdrs_plt_bar}
\title{Dichotomous Question Bar Plots.}
\usage{
cdrs_plt_bar(prep_)
}
\arguments{
\item{prep_}{the tibble returned from \code{cdrs_plt_prep.}}
}
\value{
an object of class ggplot.
}
\description{
Plot questions from the DRS data set with only one level of interest (eg. questions where we only want to show "Yes" responses).
}
