% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_hist}
\alias{cdrs_plt_hist}
\title{Prepare and plot a histogram with ggplot2.}
\usage{
cdrs_plt_hist(prep_)
}
\arguments{
\item{prep_}{list. An object returned by \code{cdrs_plt_prep}.}
}
\value{
ggplot2 object.
}
\description{
Creates a ggplot2 histogram with sample-weighted data ultimately dervied from \code{survey::svyhist}.
}
