% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_prep}
\alias{cdrs_plt_prep}
\title{Prepare data for plotting.}
\usage{
cdrs_plt_prep(
  data_,
  cols_,
  dict_ = NULL,
  level_ = NULL,
  remove_angle_brackets = TRUE,
  is_weighted = TRUE
)
}
\arguments{
\item{data_}{the DRS data.}

\item{cols_}{the columns of interest.}

\item{dict_}{the data dictionary. If \code{NULL} no plot label decoration performed. In other words, the plot will not display textual descriptions.}

\item{level_}{character. The name of the level you want to keep. This is useful for questions with dichotomous response options like \code{"Yes"}, that you wish to isolate. Should be used with \code{cdrs_plt_bar}.}

\item{remove_angle_brackets}{logical.}

\item{is_weighted}{logical.}

\item{label_details}{character.}
}
\value{
object of class tibble, data set proportions.
}
\description{
Prepare DRS data for plotting.
}
