% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrappers.R
\name{cdrs_plt_stacked}
\alias{cdrs_plt_stacked}
\title{Creates stacked bar plot.}
\usage{
cdrs_plt_stacked(data_, cols_, remove_angle_brackets = TRUE)
}
\arguments{
\item{data_}{the DRS data containing at least the column specified in the next parameter. If weighted is TRUE, both Zone and WTFINAL should also be available in data_.}

\item{cols_}{column name(s) (as a character vector)}

\item{remove_angle_brackets}{logical. Should we remove brackets around \if{html}{\out{<missingness>}} variables?}
}
\value{
an object of class ggplot.
}
\description{
Creates stacked bar plot which shows all levels of a variable. Used for both qualitative and likert scale responses.
}
