% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{cdrs_read}
\alias{cdrs_read}
\title{Read the California Delta Residents Survey Dataset}
\usage{
cdrs_read(path_ = NULL, convert_to_NA = FALSE)
}
\arguments{
\item{path_}{is a character vector of length 1. Provides the path to the directory containing all the relevant files, including the data dictionary, the data in csv form, and the hash.txt. This path may also lead to a zip file. By default the path is to the current working directory.}

\item{convert_to_NA}{is a logical vector of length 1. Converts the variety of missingness levels (eg. <Decline to answer>) to \code{NA} values, simplifying calculations. Importantly, if there are columns that were originally factors (due to different missingness levels), but is principally a numeric column, this function will adjust these columns to type numeric.}
}
\value{
Returns a tibble of the Delta Residents Survey 2023 data set.
}
\description{
Loads, validates, and corrects factor levels in the Delta Residents Survey
2023 data. Depends on cdrs_as_NA, cdrs_validate.
}
\examples{
dat <- cdrs_read(
  path_ = system.file("extdata", "demo", package = "cdrs")
)
}
