% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{cdrs_read_example}
\alias{cdrs_read_example}
\title{Load fabricated example data.}
\usage{
cdrs_read_example(convert_to_NA = FALSE)
}
\arguments{
\item{convert_to_NA_}{logical. Whether to convert to NA or not.}
}
\value{
Returns a tibble of the a fictitious data set.
}
\description{
Using cdrs::cdrs_read(), loads demo data (inside extdata/demo)
}
\examples{
demo <- cdrs_read_example(convert_to_NA = TRUE)
}
