% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{cdrs_read_example}
\alias{cdrs_read_example}
\title{Load fabricated example data.}
\usage{
cdrs_read_example(relevel_ = "default", return_dict = FALSE)
}
\arguments{
\item{relevel_}{character or list. See documentation for \code{?cdrs_read()}.}

\item{return_dict}{logical. See documentation for \code{?cdrs_read()}}
}
\value{
Returns a tibble of the a fabricated data set, or a list with the fabricated data set and the data dictionary.
}
\description{
Using cdrs::cdrs_read(), loads demo data (inside extdata/demo). Do not draw
conclusions from this fabricated "data set". The data were created by
scrambling the public version of the DRS data.
}
\examples{
demo <- cdrs_read_example()
}
