% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{remove_angle_brackets}
\alias{remove_angle_brackets}
\title{Removes angle brackets from missing variable levels.}
\usage{
remove_angle_brackets(data_, cols_ = NULL)
}
\arguments{
\item{data_}{DRS data.frame, tibble or vector.}

\item{cols_}{is a character vector (or \code{NULL}) containing name of columns of interest.}
}
\value{
the transformed DRS data set (tibble), or transformed factor/character vector.
}
\description{
Remove angle brackets around missing variable levels. This is often used for plotting functions.
}
